Update uses of checkArrayCompClassPrimitiveValueType
The utility method checkArrayCompClassPrimitiveValueType has been
renamed testIsArrayClassNullRestrictedType, to reflect more recent
terminology, and it no longer expects an "if" OpCode argument.
Instead, it generates IL that yields a zero or non-zero result to
indicate whether the array is null-restricted.  This change updates uses
of that method to branch appropriately based on the result of the IL
produced by testIsArrayClassNullRestrictedType.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>