Clean up references to old opcodes
The old opcodes had the letter i as a prefix if the
operation was indirect. Those opcodes were renamed
to use the letter i as a suffix instead.

This commit cleans up the references to the old opcodes,
isload/isstore ifload/ifstore idload/idstore to
sloadi/sstorei floadi/fstorei dloadi/dstorei

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>